добавляет контроллер фильма
